pyproject.toml
[tool.poetry]
name = "fapi"
version = "0.1.0"
description = ""
authors = ["Никита Шидогубов <nikita.sh@gklifts.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.114.0"
uvicorn = "^0.30.6"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.34"
passlib = "^1.7.4"
python-dotenv = "^1.0.1"
python-jose = "^3.3.0"
alembic = "^1.13.2"
httpx = "^0.27.2"
pytest-asyncio = "^0.24.0"
pytest = "^8.3.2"
pydantic = {extras = ["email"], version = "^2.9.0"}
pydantic-settings = "^2.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
